annotation_tool/Annotation-Interface/backend/TAPIR/requirements.txt
absl-py
notebook
jupyter_http_over_ws
chex
jax
jaxline
optax
dm-haiku
dm-tree
typing_extensions
matplotlib
mediapy
opencv-python
einshape
tensorflow
tensorflow-datasets
tensorflow_graphics
chex
jax
jaxline
optax
dm-haiku
dm-tree
typing_extensions
matplotlib
mediapy
opencv-python
einshape
ipympl
tqdm